# services/imis-connect/src/main/java/com/ozonehis/imisconnect/dto/ClaimResponse.java
package com.ozonehis.imisconnect.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ClaimResponse {

    private String claimId;

    private String status;

    private String message;
}